📦  Card-Rename
Renames Cards so they comply with the `card-{{name}}` convention rather than `{{name}}-card`
Adds a unkonwnCardHandler to the editor.
