+++ app/styles/addons/gh-koenig/gh-koenig.css
@@ -169,6 +169,6 @@ textarea.ed_code {
  * HTML Card
  */
 
- .kg-html-card {
+ .kg-card-html {
      border: #ddd 1px solid;
  }
