+++ lib/gh-koenig/addon/cards/index.js
@@ -1,7 +1,7 @@
-import htmlCard from 'gh-koenig/cards/html-card_dom';
-import imageCard from 'gh-koenig/cards/image-card_dom';
-import markdownCard from 'gh-koenig/cards/markdown-card_dom';
-import hrCard from 'gh-koenig/cards/hr-card_dom';
+import htmlCard from 'gh-koenig/cards/card-html_dom';
+import imageCard from 'gh-koenig/cards/card-image_dom';
+import markdownCard from 'gh-koenig/cards/card-markdown_dom';
+import hrCard from 'gh-koenig/cards/card-hr_dom';
 
 let cards = [];
