+++ lib/gh-koenig/addon/components/cards/card-hr.js
@@ -1,5 +1,5 @@
 import Component from 'ember-component';
-import layout from '../../templates/components/hr-card';
+import layout from '../../templates/components/card-hr';
 export default Component.extend({
     layout
 });
