+++ lib/gh-koenig/addon/templates/components/card-hr.hbs
