+++ lib/gh-koenig/addon/templates/components/card-html.hbs
