+++ lib/gh-koenig/addon/templates/components/card-image.hbs
