+++ lib/gh-koenig/addon/templates/components/card-markdown.hbs
