+++ lib/gh-koenig/app/components/card-markdown.js
@@ -1,0 +1,1 @@
+export {default} from 'gh-koenig/components/cards/card-markdown';
