--- lib/gh-koenig/app/components/hr-card.js
@@ -1,1 +1,0 @@
-export {default} from 'gh-koenig/components/cards/hr-card';
