--- lib/gh-koenig/app/components/html-card.js
@@ -1,1 +1,0 @@
-export {default} from 'gh-koenig/components/cards/html-card';
