--- lib/gh-koenig/app/components/image-card.js
@@ -1,1 +1,0 @@
-export {default} from 'gh-koenig/components/cards/image-card';
