--- lib/gh-koenig/app/components/markdown-card.js
@@ -1,1 +1,0 @@
-export {default} from 'gh-koenig/components/cards/markdown-card';
